Use u32 alignment in code gen
Code gen should ideally never depend on usize,
because that changes based on the compiler rather than for
the target

+++ crates/bindgen/src/types.rs
@@ -76,7 +76,7 @@ impl Types {
     }
 
     /// Contrast this with the size_ignoring_alignment method
-    pub fn size_rounded_to_alignment(&self, id: TypeId) -> usize {
+    pub fn size_rounded_to_alignment(&self, id: TypeId) -> u32 {
         let size_ignoring_alignment = self.size_ignoring_alignment(id);
         let alignment = self.align(id);
